collection:[#57] Store position in two dbs
Previous design was not good as it will mean value in db can be potentionally
huge, as we have few groups but many folders.
So now it's rather organized by folders.
Plus we have another tree for latest folder by group.

As we have now two trees for position operation should use transaction.